backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
jinja2
pydantic
python-multipart
pillow
